pyproject.toml
[project]
name = "sel"
version = "0.1.0"
description = "Selenium tutorial"
authors = [
    {name = "Varga Dio Jozsef", email = "31209755+cogitoergoread@users.noreply.github.com"},
]
dependencies = ["selenium>=4.32.0"]
requires-python = ">=3.13"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"


[dependency-groups]
lint = [
    "ruff>=0.11.9",
]
pytest = [
    "pytest-cov>=6.1.1",
    "pytest-selenium>=4.1.0",
]

[tool.ruff]
lint.select = [
    "B", # flake8-bugbear
    "C4", # flake8-comprehensions
    "E", # pycodestyle - Error
    "F", # Pyflakes
    "I", # isort
    "W", # pycodestyle - Warning
    "UP", # pyupgrade
]
lint.ignore = [
    "E501", # line-too-long
    "W191", # tab-indentation
]
include = ["**/*.py", "**/*.pyi", "**/pyproject.toml"]
[tool.ruff.lint.pydocstyle]
convention = "google"


[tool.pdm]
distribution = true

[tool.pytest.ini_options]
log_level = "DEBUG"
addopts = '''
--cov=sel
 --color=yes
 --cov-report=term
 --log-cli-level=DEBUG
'''
pythonpath = [
  "src/sel"
]
